+++ Info/urls.py
@@ -18,6 +18,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('',include('apps.urls')),
+    path('', include('apps.usuarios.urls')), 
     path('accounts/', include('django.contrib.auth.urls')),
 ]
