+++ Info/views.py
@@ -1,5 +1,2 @@
 from django.shortcuts import render
-from django.contrib.auth.forms import UserCreationForm
 
-def Home(request):
-	return render(request,'home.html')
