--- apps/readme.txt
@@ -1,1 +1,0 @@
-hola
