+++ templates/registration/login.html
@@ -4,16 +4,17 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login</title>
+    <title>Iniciar Sesion</title>
 </head>
 <body>
-    <h2>LOGIN</h2>
-    <form action="" method="POST">
+    <h1>INICIE SESION</h1>
+    <form action="" method="post">
+
         {% csrf_token %}
 
-        <table> {{ form.as_table }} </table>
+        {{ form }}
 
-        <input type="submit" value="Ingresar">
+        <input type="submit" value="Inicie Sesion">
     </form>
 </body>
 </html>
