+++ templates/registration/registro.html
@@ -4,14 +4,17 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>REGISTRO</title>
 </head>
 <body>
-    <form class="class-form" method="post">
-        {% csrf_token %}
-        {{ form.as_p}}
+    <h1>REGISTRARSE</h1>
+    <form action="" method="POST">
 
-        <button type="submit" id="submit">Registrarse</button>
+    {%csrf_token%}
+    {{ form.as_p}}
+
     </form>
+
+    <input type="submit" value="REGISTRARSE">
 </body>
 </html>
