--- templates/registro.html
@@ -1,7 +1,0 @@
-<form class="form" method="post">
-    {% csrf_token %}
-    {{form.as_p }}
-    
-     <button type="submit" id="submit">Registrar</button>
-    
-</form>
